tidied up sacc plotter
