+++ pipeline/sacc_plotter.py
@@ -1,124 +1,10 @@
-import matplotlib.pyplot as plt
-from soopercool import BBmeta
+from soopercool import BBmeta, ps_utils, utils
 import argparse
 from itertools import product
 import sacc
 import numpy as np
 import healpy as hp
 import os
-
-
-def load_bpwins(coupling_file):
-    """
-    """
-    bp_win = np.load(coupling_file)
-    return bp_win["bp_win"]
-
-
-def binned_theory_from_unbinned(clth, bpw_mat, lb):
-    """
-    """
-    modes = ["TT", "TE", "TB", "ET", "BT", "EE", "EB", "BE", "BB"]
-    clth_vec = np.concatenate(
-        [clth[mode] for mode in modes]
-    ).reshape(len(modes), -1)
-    _, n_bins, _, nl = bpw_mat.shape
-    nl_th = clth["EE"].shape[-1]
-    nl = min(nl, nl_th)
-    nb = len(lb)
-
-    clth_binned = np.einsum("ijkl,kl->ij", bpw_mat[:, :nb, :, :nl], clth_vec)
-    cl_out = {}
-    for i, m in enumerate(modes):
-        cl_out[m] = clth_binned[i]
-    return cl_out
-
-
-def multipole_min_from_tf(tf_file, field_pairs, snr_cut=3.):
-    """
-    """
-    tf = np.load(tf_file)
-    idx_bad_tf = {}
-    for fp in field_pairs:
-        name = f"{fp}_to_{fp}"
-        snr = tf[name] / tf[f"{name}_std"]
-        idx = np.where(snr < snr_cut)[0]
-        idx_bad_tf[fp] = idx.max() if idx.size > 0 else 0
-
-    return idx_bad_tf
-
-
-def plot_spectrum(lb, cb, cb_err, title, ylabel, xlim,
-                  cb_data=None, cb_data_err=None, add_theory=False,
-                  lth=None, clth=None, cbth=None, save_file=None):
-    """
-    """
-    plt.figure(figsize=(8, 6))
-    grid = plt.GridSpec(4, 1, wspace=0, hspace=0)
-    if add_theory:
-        main = plt.subplot(grid[:-1])
-        sub = plt.subplot(grid[-1])
-        sub.set_xlabel(r"$\ell$")
-        sub.set_ylabel(r"$\Delta C_\ell / (\sigma / \sqrt{N_\mathrm{sims}})$")
-    else:
-        main = plt.subplot(grid[:])
-        main.set_xlabel(r"$\ell$")
-
-    main.set_ylabel(r"$\ell (\ell + 1) C_\ell^{%s} / 2\pi$" % ylabel)
-    main.set_title(title)
-
-    fac = lb * (lb + 1) / (2 * np.pi)
-    offset = 0 if cb_data is None else 1
-
-    if add_theory:
-        fac_th = lth * (lth + 1) / (2 * np.pi)
-        main.plot(lth, fac_th * clth, c="darkgray", ls="-.", lw=2.6,
-                  label="theory")
-
-    main.errorbar(
-        lb-offset, fac * cb, yerr=fac * cb_err, marker="o", ls="None",
-        markerfacecolor="white", markeredgecolor="navy", label="sims",
-        elinewidth=1.75, ecolor="navy", markeredgewidth=1.75
-    )
-
-    if cb_data is not None:
-        main.errorbar(
-            lb+offset, fac * cb_data, yerr=fac * cb_data_err, marker="o",
-            ls="None", markerfacecolor="white", markeredgecolor="darkorange",
-            elinewidth=1.75, ecolor="darkorange", markeredgewidth=1.75,
-            label="data"
-        )
-
-    main.legend(fontsize=13)
-    main.set_xlim(*xlim)
-
-    if add_theory:
-        sub.axhspan(-3, 3, color="gray", alpha=0.1)
-        sub.axhspan(-2, 2, color="gray", alpha=0.4)
-        sub.axhspan(-1, 1, color="gray", alpha=0.8)
-
-        sub.plot(
-            lb-offset, (cb - cbth) / cb_err, marker="o", ls="None",
-            markerfacecolor="white", markeredgecolor="navy",
-            markeredgewidth=1.75
-        )
-        if cb_data is not None:
-            sub.plot(
-                lb+offset, (cb_data - cbth) / cb_data_err,
-                marker="o", ls="None",
-                markerfacecolor="white", markeredgecolor="darkorange",
-                markeredgewidth=1.75
-            )
-
-        sub.set_xlim(*xlim)
-        sub.set_ylim(-4.5, 4.5)
-
-    if save_file:
-        plt.savefig(save_file, bbox_inches="tight")
-    else:
-        plt.tight_layout()
-        plt.show()
-    plt.close()
 
 
 def main(args):
@@ -132,16 +18,8 @@ def main(args):
 
     out_dir = meta.output_directory
     sacc_dir = f"{out_dir}/saccs"
-    couplings_dir = f"{out_dir}/couplings"
-    if "couplings_directory" in meta.covariance:
-        if meta.covariance["couplings_directory"] is not None:
-            couplings_dir = meta.covariance["couplings_directory"]
-
     plot_dir = f"{out_dir}/plots/sacc_spectra"
     BBmeta.make_dir(plot_dir)
-
-    nmt_binning = meta.read_nmt_binning()
-    lb = nmt_binning.get_effective_ells()
     lmax = meta.lmax
 
     beams = {
@@ -165,18 +43,18 @@ def main(args):
         tf_dir = f"{transfer_dir}/transfer_function_{ftag1}_x_{ftag2}.npz"
         if not os.path.isfile(tf_dir):
             tf_dir = f"{transfer_dir}/transfer_function_{ftag2}_x_{ftag1}.npz"
-        idx = multipole_min_from_tf(
+        idx = utils.multipole_min_from_tf(
             tf_dir,
             field_pairs=field_pairs,
             snr_cut=3
         )
         idx_bad_tf[ftag1, ftag2] = idx
 
+    bpws = meta.get_inverse_couplings(return_bpwf=True)[1]
     for ms1, ms2 in ps_names:
-        bpw_file = f"{couplings_dir}/couplings_{ms1}_{ms2}.npz"
-        if not os.path.isfile(bpw_file):
-            bpw_file = f"{couplings_dir}/couplings_{ms2}_{ms1}.npz"
-        bpw_mats[ms1, ms2] = load_bpwins(bpw_file)
+        ftag1, ftag2 = (meta.filtering_tag_from_map_set(ms)
+                        for ms in (ms1, ms2))
+        bpw_mats[ms1, ms2] = bpws[ftag1, ftag2]
 
     plot_sims = {
         (ms1, ms2, fp): {
@@ -229,8 +107,7 @@ def main(args):
                 # reverse the order of the two fields
                 cl_th[fp] = clth[fp[::-1]]
 
-        clth_binned = binned_theory_from_unbinned(cl_th, bpw_mats[(ms1, ms2)],
-                                                  lb[lb <= meta.lmax])
+        clth_binned = ps_utils.bin_theory_cls(cl_th, bpw_mats[(ms1, ms2)])
 
         ftag1 = meta.filtering_tag_from_map_set(ms1)
         ftag2 = meta.filtering_tag_from_map_set(ms2)
@@ -328,7 +205,7 @@ def main(args):
             plot_sims[ms1, ms2, fp]["err"] /= np.sqrt(Nsims)
             plot_name = f"plot_cells_{ms1}_{ms2}_{fp}.pdf"
 
-            plot_spectrum(
+            ps_utils.plot_spectrum(
                 plot_sims[ms1, ms2, fp]["x"],
                 plot_sims[ms1, ms2, fp]["y"],
                 plot_sims[ms1, ms2, fp]["err"],
